Cleanup Interface::set_project()
Just delete all sheet widgets when setting a new project. Although
comments say it's probably not safe to do this, it seems to work
reliably now, so it must be a problem from the past.

It fixes the problem of not having input focus when loading a new
project without closing the old one explicitly.