cleanup keymap a bit
move_left/right slots for move commands don't need to be added for each
and every command that inherits MoveCommand.
So added it just for MoveCommand now.
It is possible though to define it also for a command class that
inherits MoveCommand to override the autorepeat variables from the
MoveCommand one!