update clips read source active state if clip was added / removed
from it's track.

If the clip is removed from it's track, the readsource no longer has to
be kept in sync.
If the clip is added to it's track, then initially set it's read source
active, but then also check if it has to be active in case it's track is
muted or muted by solo or ...