support inheritance for delegated keybindings.
E.g. a key binding is mapped to an object and it's slot, but you want
the key to be delegated to another context item.

Say, object is TrackPanelView, and below TrackPanelView is TrackView.
Now when left clicking with the mouse on the TrackPanelView I want to
invoke TrackView::edit_properties() but not when left clicking the mouse
on the track view itself.

Now you can do so by delegating left mouse button to TrackView:
<objectname="TrackPanelView"slotsignature="TrackView::edit_properties"/>

Rationale: you can bind a key binding to an object A that partially
covers another object B, and you want this B's object's edit_properties
function called only when context is on object A.

Or something like that ;D