manage 'software' and 'harware' audiobuses in Project
Let Project manage all the AudioBuses and AudioChannels

If Jack is the driver, give a list of channels we want to the
AudioDevice, so JackDriver can create the ports we want.

if Jack is not the driver, so we have a real hardware driver, then let
AudioDevice init it's driver, and let Project wrap the available
AudioChannels into AudioBuses, also managed by Project.

With these changes it has become much easier to store and restore the
audio i/o configuration on the hardware and software level, even if the
'wrong' driver was loaded for a certain project.