changed the appearance of the correlation meter
In some situations the correlation meter was hard to read. Namely when
the foreground gradient faded into the background colour, it was hard to
see where the correlation range ends. With the new pointy shape this
problem is solved. I implemented 2 versions, one with cubic splines,
which is a horrible resource hog, and one with straight lines, which
looks uglier. Change the bool "useCubicSpline" to false (in
CorrelationMeterWidget.cpp) if you prefer the more efficient straight
lines.