first round of unifying canvas and hold cursor
by removing support of the hardware cursor on the canvas (setting it to
a blank cursor)
It uses a little more cpu perhaps, but moving the mouse causes canvas
items to highlight anyway, consuming about the same amount of cpu, so to
me it's a non-issue.

Upside: it removes the visible lag when the hardware cursor is moved
dragging for example the work cursor with it. Much more pleasing to the
eye ;)